+++ CHANGELOG.md
@@ -9,14 +9,34 @@ This changelog follows the principles of [Keep a Changelog](https://keepachangel
 ### Added
 
 - Collections: Added theme information when retrieving a collection using `getCollection`.
+
+### Changed
+
+### Fixed
+
+### Removed
+
+## [v2.2.0] -- 2026-04-24
+
+### Added
+
 - Datasets: Added `updateDatasetLicense` use case and repository method to support Dataverse endpoint `PUT /datasets/{id}/license`, for updating dataset license or custom terms.
 - Datasets: Added `getDatasetStorageDriver` use case and repository method to support Dataverse endpoint `GET /datasets/{identifier}/storageDriver`, for retrieving dataset storage driver configuration with properties: name, type, label, directUpload, directDownload, and uploadOutOfBand.
-- Datasets: Added `updateDatasetLicense` use case and repository method to support Dataverse endpoint `PUT /datasets/{id}/license`, for updating dataset license or custom terms
+- Datasets: Added `getDatasetUploadLimits` use case and repository method to support Dataverse endpoint `GET /datasets/{id}/uploadlimits`, for retrieving remaining storage upload quotas, if present.
 - New Use Case: [Get Collections For Linking Use Case](./docs/useCases.md#get-collections-for-linking).
 - New Use Case: [Create a Template](./docs/useCases.md#create-a-template) under Templates.
 - New Use Case: [Get a Template](./docs/useCases.md#get-a-template) under Templates.
 - New Use Case: [Delete a Template](./docs/useCases.md#delete-a-template) under Templates.
+- Templates: Added `setTemplateAsDefault` use case and repository method to support Dataverse endpoint `POST /dataverses/{id}/template/default/{templateId}`.
+- Templates: Added `unsetTemplateAsDefault` use case and repository method to support Dataverse endpoint `DELETE /dataverses/{id}/template/default`.
 - New Use Case: [Update Terms of Access](./docs/useCases.md#update-terms-of-access).
+- Guestbooks: Added use cases and repository support for guestbook creation, listing, and enabling/disabling.
+- Guestbooks: Added dataset-level guestbook assignment and removal support via `assignDatasetGuestbook` (`PUT /api/datasets/{identifier}/guestbook`) and `removeDatasetGuestbook` (`DELETE /api/datasets/{identifier}/guestbook`).
+- Datasets/Guestbooks: Added `guestbookId` in `getDataset` responses.
+- Access: Added`access` module for guestbook-at-request and download terms/guestbook submission endpoints.
+- New Use Case: [Get Publish Dataset Disclaimer Text](./docs/useCases.md#get-publish-dataset-disclaimer-text).
+- New Use Case: [Get Dataset Publish Popup Custom Text](./docs/useCases.md#get-dataset-publish-popup-custom-text).
+- DatasetType: Updated datasetType data model. Added two more fields: description and displayName.
 
 ### Changed
 
@@ -34,7 +54,7 @@ This changelog follows the principles of [Keep a Changelog](https://keepachangel
 
 - Removed date fields validations in create and update dataset use cases, since validation is already handled in the backend and SPA frontend (other clients should perform client side validation also). This avoids duplicated logic and keeps the package focused on its core responsibility.
 
-[Unreleased]: https://github.com/IQSS/dataverse-client-javascript/compare/v2.1.0...develop
+[Unreleased]: https://github.com/IQSS/dataverse-client-javascript/compare/v2.2.0...develop
 
 ---